+++ .htaccess
@@ -1,0 +1,1 @@
+DirectoryIndex www/
