+++ descqa/__init__.py
@@ -1,1 +1,3 @@
 # This is an empty file to make this directory a python package
+from .register import *
+__version__ = '2.4.0'
