+++ descqa/configs/ListQuantities.yaml
@@ -1,0 +1,1 @@
+subclass_name: ListAvailableQuantities.ListAvailableQuantities
