+++ descqa/configs/SkyArea.yaml
@@ -1,0 +1,2 @@
+subclass_name: SkyArea.SkyArea
+nside: 16
