+++ descqa/utils.py
