+++ descqarun/archiver.py
@@ -5,11 +5,15 @@ import time
 import subprocess
 import argparse
 
+
+__all__ = ['archive', 'main']
+
+
 def archive(src, dst):
     if os.path.exists(dst):
         raise ValueError('{} already exists'.format(dst))
     subprocess.check_call(['tar', '-czf', dst, src])
-    subprocess.check_call(['rm', '-rf', src])    
+    subprocess.check_call(['rm', '-rf', src])
 
 
 def main():
@@ -18,7 +22,7 @@ def main():
     parser.add_argument('dest_dir')
     parser.add_argument('archive_to_date', help='in the format of YYYY-MM-DD')
     args = parser.parse_args()
-    
+
     if re.match(r'20\d{2}-[01]\d-[0123]\d', args.archive_to_date) is None:
         parser.error('`archive_to_date` must be in the format of YYYY-MM-DD')
 
@@ -26,7 +30,7 @@ def main():
     for d in items:
         src = os.path.join(args.source_dir, d)
         dst = os.path.join(args.dest_dir, d + '.tar.gz')
-        
+
         if not os.path.isdir(src):
             continue
         m = re.match(r'(20\d{2}-[01]\d-[0123]\d)(?:_\d+)?', d)
@@ -45,4 +49,3 @@ def main():
 
 if __name__ == '__main__':
     main()
-
