+++ descqaweb/templates/leftpanel.html
