+++ descqaweb/twopanels.py
