--- index.html
@@ -1,7 +1,0 @@
-<!DOCTYPE HTML>
-<html>
-  <head>
-    <meta http-equiv="refresh" content="0; url=www/">
-  </head>
-</html>
-
