--- validation_code/README.md
@@ -1,7 +1,0 @@
-# DESCQA Validation Tests
-
-This directory contains code to conduct validation tests.
-
-## How to create a new test
-
-Under construction.
