--- validation_configs/CLF_r.yaml
@@ -1,4 +1,0 @@
-subclass_name: ConditionalLuminosityFunction
-band2: r
-band1: g
-color_cut_fraction: 0.2
