--- validation_configs/ListQuantities.yaml
@@ -1,1 +1,0 @@
-subclass_name: ListAvailableQuantities
