--- validation_configs/SkyArea.yaml
@@ -1,2 +1,0 @@
-subclass_name: SkyArea
-nside: 16
